requirements.txt
openai
langchain
python-dotenv
streamlit==1.39.0
pandas==2.2.2
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0